tenders: Share common code
Introduce a tenders/common/libcommon.a, and share common code between
the hvt and spt tender. This includes, at the moment:

- the ELF loader, renamed to just elf_load()
- common parts of block/net setup code, now in tap_attach() and
block_attach()

While we're at it, replace the unsafe use of atoi() in tap_attach() with
strtol() and properly range-check the returned value.
