+++ .gitignore
@@ -1,5 +1,6 @@
 *.o
 *.d
+*.a
 *.lib.so
 *.pc
 Makeconf
