+++ tenders/GNUmakefile
@@ -28,54 +28,95 @@ all_TARGETS :=
 .SUFFIXES:
 $(V).SILENT:	
 
-ifdef CONFIG_HVT
+ifneq ($(filter 1,$(CONFIG_HVT) $(CONFIG_SPT)),)
 
-hvt_SRCS := hvt_core.c hvt_elf.c hvt_main.c hvt_cpu_$(CONFIG_ARCH).c
-hvt_MODULES ?= blk net
+common_LIB := common/libcommon.a
+common_SRCS := common/elf.c common/block_attach.c common/tap_attach.c
+common_OBJS := $(patsubst %.c,%.o,$(common_SRCS))
 
-ifeq ($(CONFIG_HOST), Linux)
-    hvt_SRCS += hvt_kvm.c hvt_kvm_$(CONFIG_ARCH).c
-    hvt_debug_MODULES ?= gdb dumpcore
-    all_TARGETS += solo5-hvt solo5-hvt-debug
+$(common_LIB): $(common_OBJS)
+	@echo HOSTAR $@
+	$(HOSTAR) rc $@ $^
 
-    HOSTLDFLAGS += -Wl,-z -Wl,noexecstack
-else ifeq ($(CONFIG_HOST), FreeBSD)
-    hvt_SRCS += hvt_freebsd.c hvt_freebsd_$(CONFIG_ARCH).c
-    hvt_debug_MODULES ?= gdb dumpcore
-    all_TARGETS += solo5-hvt solo5-hvt-debug
-else ifeq ($(CONFIG_HOST), OpenBSD)
-    hvt_SRCS += hvt_openbsd.c hvt_openbsd_$(CONFIG_ARCH).c
-    all_TARGETS += solo5-hvt
-endif
-
-hvt_SRCS += $(patsubst %,hvt_module_%.c,$(hvt_MODULES))
-hvt_debug_SRCS += $(patsubst %,hvt_module_%.c,$(hvt_debug_MODULES))
-hvt_OBJS := $(patsubst %.c,%.o,$(hvt_SRCS))
-hvt_debug_OBJS := $(patsubst %.c,%.o,$(hvt_debug_SRCS))
+all_TARGETS += $(common_LIB)
 
 %.o: %.c %.d
 	$(HOSTCOMPILE.c)
 
+%.o: %.S %.d
+	$(HOSTCOMPILE.S)
+
 %.d: ;
 .PRECIOUS: %.d
 
-solo5-hvt: $(hvt_OBJS)
+endif
+
+ifdef CONFIG_HVT
+
+hvt_SRCS := hvt/hvt_core.c hvt/hvt_main.c hvt/hvt_cpu_$(CONFIG_ARCH).c
+hvt_MODULES ?= blk net
+
+ifeq ($(CONFIG_HOST), Linux)
+    hvt_SRCS += hvt/hvt_kvm.c hvt/hvt_kvm_$(CONFIG_ARCH).c
+    hvt_debug_MODULES ?= gdb dumpcore
+    all_TARGETS += hvt/solo5-hvt hvt/solo5-hvt-debug
+
+    HOSTLDFLAGS += -Wl,-z -Wl,noexecstack
+else ifeq ($(CONFIG_HOST), FreeBSD)
+    hvt_SRCS += hvt/hvt_freebsd.c hvt/hvt_freebsd_$(CONFIG_ARCH).c
+    hvt_debug_MODULES ?= gdb dumpcore
+    all_TARGETS += hvt/solo5-hvt hvt/solo5-hvt-debug
+else ifeq ($(CONFIG_HOST), OpenBSD)
+    hvt_SRCS += hvt/hvt_openbsd.c hvt/hvt_openbsd_$(CONFIG_ARCH).c
+    all_TARGETS += hvt/solo5-hvt
+endif
+
+hvt_SRCS += $(patsubst %,hvt/hvt_module_%.c,$(hvt_MODULES))
+hvt_debug_SRCS += $(patsubst %,hvt/hvt_module_%.c,$(hvt_debug_MODULES))
+hvt_OBJS := $(patsubst %.c,%.o,$(hvt_SRCS))
+hvt_debug_OBJS := $(patsubst %.c,%.o,$(hvt_debug_SRCS))
+
+hvt/solo5-hvt: $(hvt_OBJS) $(common_LIB)
 	$(HOSTLINK)
 
-solo5-hvt-debug: $(hvt_OBJS) $(hvt_debug_OBJS)
+hvt/solo5-hvt-debug: $(hvt_OBJS) $(hvt_debug_OBJS) $(common_LIB)
 	$(HOSTLINK)
 
 endif # CONFIG_HVT
 
+ifdef CONFIG_SPT
+
+HOSTLDFLAGS += -Wl,-z -Wl,noexecstack
+
+ifdef CONFIG_SPT_NO_PIE
+
+HOSTLDFLAGS += -Wl,-Ttext-segment=0x40000000
+
+endif
+
+HOSTLDLIBS += -lseccomp
+
+spt_SRCS := spt/spt_main.c spt/spt_core.c spt/spt_launch_$(CONFIG_ARCH).S \
+    spt/spt_module_net.c spt/spt_module_block.c
+
+spt_OBJS := $(patsubst %.c,%.o,$(patsubst %.S,%.o,$(spt_SRCS)))
+
+spt/solo5-spt: $(spt_OBJS) $(common_LIB)
+	$(HOSTLINK)
+
+all_TARGETS += spt/solo5-spt
+
+endif # CONFIG_SPT
+
 all: $(all_TARGETS)
 
-all_OBJS := $(hvt_OBJS) $(hvt_debug_OBJS)
+all_OBJS := $(common_OBJS) $(hvt_OBJS) $(hvt_debug_OBJS) $(spt_OBJS)
 all_DEPS := $(patsubst %.o,%.d,$(all_OBJS))
 
 .PHONY: clean
 
 clean:
-	@echo "CLEAN tenders/hvt"
+	@echo "CLEAN tenders"
 	$(RM) $(all_TARGETS) $(all_OBJS) $(all_DEPS)
 
 include $(wildcard $(all_DEPS))
