+++ tenders/common/cc.h
