+++ tenders/hvt/hvt_main.c
@@ -191,7 +191,7 @@ int main(int argc, char **argv)
     hvt_mem_size(&mem_size);
     struct hvt *hvt = hvt_init(mem_size);
 
-    hvt_elf_load(elffile, hvt->mem, hvt->mem_size, &gpa_ep, &gpa_kend);
+    elf_load(elffile, hvt->mem, hvt->mem_size, &gpa_ep, &gpa_kend);
 
     char *cmdline;
     hvt_vcpu_init(hvt, gpa_ep, gpa_kend, &cmdline);
