+++ tenders/hvt/hvt_module_net.c
@@ -25,171 +25,19 @@
 #include <assert.h>
 #include <err.h>
 #include <errno.h>
-#include <fcntl.h>
-#include <ifaddrs.h>
 #include <inttypes.h>
 #include <stdint.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
-#include <sys/ioctl.h>
 #include <unistd.h>
 
+#include "../common/tap_attach.h"
 #include "hvt.h"
-
-#if defined(__linux__)
-
-/*
- * Linux TAP device specific.
- */
-#include <sys/socket.h>
-#include <linux/if.h>
-#include <linux/if_tun.h>
-
-#elif defined(__FreeBSD__)
-
-#include <net/if.h>
-
-#elif defined(__OpenBSD__)
-
-#include <sys/socket.h>
-#include <net/if.h>
-
-#else /* !__linux__ && !__FreeBSD__ && !__OpenBSD__ */
-
-#error Unsupported target
-
-#endif
 
 static char *netiface;
 static int netfd;
 static struct hvt_netinfo netinfo;
 static int cmdline_mac = 0;
-
-/*
- * Attach to an existing TAP interface named 'ifname'.
- *
- * Returns -1 and an appropriate errno on failure (ENOENT if the interface does
- * not exist), and the tap device file descriptor on success.
- */
-static int tap_attach(const char *ifname)
-{
-    int fd;
-
-    /*
-     * Syntax @<number> indicates a pre-existing open fd, so just pass it
-     * through without any checks.
-     */
-    if (ifname[0] == '@') {
-        fd = atoi(&ifname[1]);
-
-        if (fcntl(fd, F_SETFL, O_NONBLOCK) == -1)
-            return -1;
-
-        return fd;
-    }
-
-    /*
-     * Verify that the interface exists and is up and running. If we don't do
-     * this then we get "create on open" behaviour on most systems which is not
-     * what we want.
-     */
-    struct ifaddrs *ifa, *ifp;
-    int found = 0;
-    int up = 0;
-
-    if (getifaddrs(&ifa) == -1)
-        return -1;
-    ifp = ifa;
-    while (ifp) {
-        if (strcmp(ifp->ifa_name, ifname) == 0) {
-            found = 1;
-            up = ifp->ifa_flags & (IFF_UP | IFF_RUNNING);
-            break;
-        }
-        ifp = ifp->ifa_next;
-    }
-    freeifaddrs(ifa);
-    if (!found) {
-        errno = ENOENT;
-        return -1;
-    }
-
-#if defined(__linux__)
-
-    if (!up) {
-        errno = ENETDOWN;
-        return -1;
-    }
-
-    int err;
-    struct ifreq ifr;
-
-    fd = open("/dev/net/tun", O_RDWR | O_NONBLOCK);
-    if (fd == -1)
-        return -1;
-
-    /*
-     * Initialise ifr for TAP interface.
-     */
-    memset(&ifr, 0, sizeof(ifr));
-    /*
-     * TODO: IFF_NO_PI may silently truncate packets on read().
-     */
-    ifr.ifr_flags = IFF_TAP | IFF_NO_PI;
-    if (strlen(ifname) > IFNAMSIZ) {
-        errno = EINVAL;
-        return -1;
-    }
-    strncpy(ifr.ifr_name, ifname, IFNAMSIZ);
-
-    /*
-     * Attach to the tap device; we have already verified that it exists, but
-     * see below.
-     */
-    if (ioctl(fd, TUNSETIFF, (void *)&ifr) == -1) {
-        err = errno;
-        close(fd);
-        errno = err;
-        return -1;
-    }
-    /*
-     * If we got back a different device than the one requested, e.g. because
-     * the caller mistakenly passed in '%d' (yes, that's really in the Linux
-     * API) then fail.
-     */
-    if (strncmp(ifr.ifr_name, ifname, IFNAMSIZ) != 0) {
-        close(fd);
-        errno = EINVAL;
-        return -1;
-    }
-
-#elif defined(__FreeBSD__)
-
-    char devname[strlen(ifname) + 6];
-
-    snprintf(devname, sizeof devname, "/dev/%s", ifname);
-    fd = open(devname, O_RDWR | O_NONBLOCK);
-    if (fd == -1)
-        return -1;
-
-#elif defined(__OpenBSD__)
-
-    if (!up) {
-        errno = ENETDOWN;
-        return -1;
-    }
-
-    char devname[strlen(ifname) + 6];
-
-    snprintf(devname, sizeof devname, "/dev/%s", ifname);
-    fd = open(devname, O_RDWR | O_NONBLOCK);
-    if (fd == -1)
-        return -1;
-#endif
-
-    return fd;
-}
 
 static void hypercall_netinfo(struct hvt *hvt, hvt_gpa_t gpa)
 {
@@ -255,30 +103,12 @@ static int setup(struct hvt *hvt)
     if (netiface == NULL)
         return 0; /* Not present */
 
-    /* attach to requested tap interface */
     netfd = tap_attach(netiface);
-    if (netfd < 0) {
+    if (netfd < 0)
         err(1, "Could not attach interface: %s", netiface);
-        exit(1);
-    }
 
-    if (!cmdline_mac) {
-        /* generate a random, locally-administered and unicast MAC address */
-        int rfd = open("/dev/urandom", O_RDONLY);
-
-        if (rfd == -1)
-            err(1, "Could not open /dev/urandom");
-
-        uint8_t guest_mac[6];
-        int ret;
-
-        ret = read(rfd, guest_mac, sizeof(guest_mac));
-        assert(ret == sizeof(guest_mac));
-        close(rfd);
-        guest_mac[0] &= 0xfe;
-        guest_mac[0] |= 0x02;
-        memcpy(netinfo.mac_address, guest_mac, sizeof netinfo.mac_address);
-    }
+    if (!cmdline_mac)
+	tap_attach_genmac(netinfo.mac_address);
 
     assert(hvt_core_register_hypercall(HVT_HYPERCALL_NETINFO,
                 hypercall_netinfo) == 0);
